Refactor Storage into him package
